Overhaul the list of programs I use
I'm not quite sure what started me on doing this, but it's long overdue.

Of special note is my removal of Apache Solr from the list. While I have
written software that used it, that software was never used by anyone
for real work, so I actually have no grounds for saying Solr was fit for
purpose.