Note the existence of Timezone Boundary Builder
I don't know that I'll ever actually have a use for it, but my deep
fondness for timezone trivia means it's too awesome a thing to ignore.

And to be fair, if I ever have to write code that cares about timezones
again, I actually may have a use for it.